requirements.txt
# Python dependencies for whale_worker
# Minimal dependencies for the whale trades alert worker

# HTTP client for API requests
httpx>=0.27.0

# MongoDB driver
pymongo>=4.6.0

# Environment variable loading
python-dotenv>=1.0.0

